- deleted collapsible component - updated bottomPanel to always render but display hidden
