--- ui/src/components/collapsiblePanels/components/ToggleArea.tsx
@@ -1,48 +1,0 @@
-import {
-  ChevronDownIcon,
-  ChevronUpIcon,
-  ChevronLeftIcon,
-  ChevronRightIcon,
-} from "@radix-ui/react-icons";
-
-import { IconButton } from "../..";
-
-export type ArrowPosition = "start" | "center" | "end";
-export type ArrowDirection = "left" | "right" | "up" | "down";
-
-export type ToggleProps = {
-  className?: string;
-  buttonClassName?: string;
-  arrowPosition?: ArrowPosition;
-  arrowDirection?: ArrowDirection;
-  onClick?: () => void;
-};
-
-const ToggleArea: React.FC<ToggleProps> = ({
-  className,
-  buttonClassName,
-  arrowPosition,
-  arrowDirection,
-  onClick,
-}) => (
-  <div
-    className={`flex ${arrowPosition === "end" ? "justify-end" : arrowPosition === "start" ? "justify-start" : "justify-center"} w-fill cursor-pointer ${className}`}
-    onClick={onClick}>
-    <IconButton
-      className={buttonClassName}
-      icon={
-        arrowDirection === "right" ? (
-          <ChevronRightIcon />
-        ) : arrowDirection === "up" ? (
-          <ChevronUpIcon />
-        ) : arrowDirection === "down" ? (
-          <ChevronDownIcon />
-        ) : (
-          <ChevronLeftIcon />
-        )
-      }
-    />
-  </div>
-);
-
-export { ToggleArea };
