--- ui/src/components/collapsiblePanels/index.ts
@@ -1,3 +1,0 @@
-export * from "./HorizontalPanel";
-
-export * from "./types";
