--- ui/src/components/collapsiblePanels/types.ts
@@ -1,7 +1,0 @@
-export type PanelContent = {
-  id: string;
-  component: React.ReactNode;
-  title?: string;
-  description?: string;
-  icon?: React.ReactNode;
-};
