+++ ui/src/features/BottomPanel/index.tsx
@@ -1,14 +1,20 @@
-import { HorizontalPanel, OutputIcon, PreviewIcon, type PanelContent } from "@flow/components";
+import { useCallback, useState } from "react";
+
+import { OutputIcon, PreviewIcon, IconButton } from "@flow/components";
 import { useStateManager } from "@flow/hooks";
 import { useT } from "@flow/providers";
 
 import { DataTable, LogConsole, Map } from "./components";
 
-export type BottomPanelProps = {
-  className?: string;
+type PanelContent = {
+  id: string;
+  component: React.ReactNode;
+  title?: string;
+  description?: string;
+  icon?: React.ReactNode;
 };
 
-const BottomPanel: React.FC<BottomPanelProps> = ({ className }) => {
+const BottomPanel: React.FC = () => {
   const [isPanelOpen, handlePanelToggle] = useStateManager(false);
   const t = useT();
 
@@ -32,13 +38,54 @@ const BottomPanel: React.FC<BottomPanelProps> = ({ className }) => {
     },
   ];
 
+  const [selected, setSelected] = useState<PanelContent | undefined>(panelContents?.[0]);
+
+  const baseClasses = "flex flex-col box-content transition-width duration-300 ease-in-out";
+  const classes = [
+    baseClasses,
+    isPanelOpen ? "h-100" : "h-[36px]",
+    "bg-zinc-900 border-t border-zinc-700 backdrop-blur-md",
+  ].reduce((acc, cur) => (cur ? `${acc} ${cur}` : acc));
+
+  const handleSelection = useCallback(
+    (content: PanelContent) => {
+      if (content.id !== selected?.id) {
+        setSelected(content);
+        if (!isPanelOpen) {
+          handlePanelToggle?.(true);
+        }
+      } else {
+        handlePanelToggle?.(!isPanelOpen);
+        if (content.id === selected?.id) {
+          setSelected(undefined);
+        }
+      }
+    },
+    [isPanelOpen, handlePanelToggle, selected],
+  );
+
   return (
-    <HorizontalPanel
-      className={`bg-zinc-900 border-t border-zinc-700 backdrop-blur-md ${className}`}
-      isOpen={!!isPanelOpen}
-      panelContents={panelContents}
-      onToggle={handlePanelToggle}
-    />
+    <div className={classes}>
+      <div className="flex gap-1 items-center justify-center h-[36px]">
+        {panelContents?.map(content => (
+          <IconButton
+            key={content.id}
+            className={`w-[55px] h-[80%] ${selected?.id === content.id ? "text-white bg-zinc-800" : undefined}`}
+            icon={content.icon}
+            tooltipText={content.description}
+            tooltipPosition="top"
+            onClick={() => handleSelection(content)}
+          />
+        ))}
+      </div>
+      <div id="content" className={`flex flex-1 bg-zinc-800}`}>
+        {panelContents.map(p => (
+          <div className={`flex-1 p-1 ${selected?.id === p.id ? "flex" : "hidden"}`} key={p.id}>
+            {p.component}
+          </div>
+        ))}
+      </div>
+    </div>
   );
 };
